LayoutTests/accessibility/link-inside-button-accessible-text.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test-pre.js"></script>
</head>
<body id="body">

<div id="content">

<div role="button" id="button1">
    <a href="#">foo1</a>
</div>

<button id="button2">
    <a href="#">foo2</a>
</button>

</div>

<div id="console"></div>

<script>

    description("This tests that when a link is inside a button, its visible text is still available for the button.");

    if (window.accessibilityController) {

          var button1 = accessibilityController.accessibleElementById("button1");
          shouldBe("button1.title", "'AXTitle: foo1'");

          var button2 = accessibilityController.accessibleElementById("button2");
          shouldBe("button2.title", "'AXTitle: foo2'");

          document.getElementById("content").style.visibility = 'hidden';
    }

</script>

<script src="../resources/js-test-post.js"></script>
</body>
</html>
